pyproject.toml
[project]
name = "sircyl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "authzclient>=0.1.0",
    "cryptography>=46.0.3",
    "dependency-injector>=4.48.2",
    "flask>=3.1.2",
    "logging-gelf>=0.0.32",
    "nltk==3.5",
    "pytesseract==0.3.13",
    "pdf2image==1.17.0",
    "Pillow==12.0.0",
    "pika>=1.3.2",
    "pyjwt>=2.10.1",
    "sircylclient==0.1.0a1",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[[tool.uv.index]]
name = "jcyl-gss"
url = "https://despliegues.ae.jcyl.es/artifactory/api/pypi/jcyl-gss-pypi-local/simple"
publish-url = "https://despliegues.ae.jcyl.es/artifactory/api/pypi/jcyl-gss-pypi-local"
explicit = true

[tool.uv.sources]
authzclient = { path = "../authzclient", editable=false }
sircylclient = { path = "../sircylclient", editable=true }

[dependency-groups]
dev = [
    { include-group = "lint" },
    { include-group = "test" }
]
lint = [
    "ruff>=0.14.5",
]
prod = [
]
test = [
    "pytest>=9.0.1",
]

[tool.uv]
default-groups = ["dev", "prod"]
